+++ sktime/experiments/analysis.py
@@ -2,16 +2,14 @@
 import pandas as pd
 import numpy as np
 import itertools
-
 from scipy import stats
 from scipy.stats import ttest_ind
 from scipy.stats import ranksums
-
 import scikit_posthocs as sp
-
-from sktime.experiments.losses import Losses
-
 import matplotlib.pyplot as plt
+
+from .losses import Losses
+
 
 class AnalyseResults(object):
     """
