+++ sktime/experiments/data.py
@@ -1,18 +1,21 @@
-import os
-from sktime.utils.load_data import load_from_tsfile_to_dataframe
-from sktime.utils.results_writing import write_results_to_uea_format
-from sktime.highlevel import TSCTask, ForecastingTask
-import re
-import pandas as pd
-from abc import ABC, abstractmethod
-import numpy as np
 import csv
+import os
+from abc import ABC, abstractmethod
+
+import numpy as np
+import pandas as pd
+
+from ..utils.load_data import load_from_tsfile_to_dataframe
+from ..utils.results_writing import write_results_to_uea_format
+
 
 class DatasetHDD:
     """
     Another class for holding the data
     """
-    def __init__(self, dataset_loc, dataset_name, train_test_exists=True, sufix_train='_TRAIN.ts', suffix_test='_TEST.ts' ,target='target'):
+
+    def __init__(self, dataset_loc, dataset_name, train_test_exists=True, sufix_train='_TRAIN.ts',
+                 suffix_test='_TEST.ts', target='target'):
         """
         Parameters
         ----------
@@ -51,12 +54,12 @@ class DatasetHDD:
         pandas DataFrame:
             dataset in pandas DataFrame format
         """
-        #TODO curently only the current use case with saved datasets on the disk in a certain format is supported. This should be made more general.
+        # TODO curently only the current use case with saved datasets on the disk in a certain format is supported. This should be made more general.
         if self._train_test_exists:
-   
-
-            loaded_dts_train = load_from_tsfile_to_dataframe(os.path.join(self._dataset_loc, self._dataset_name + self._suffix_train))
-            loaded_dts_test = load_from_tsfile_to_dataframe(os.path.join(self._dataset_loc, self._dataset_name + self._suffix_test))
+            loaded_dts_train = load_from_tsfile_to_dataframe(
+                os.path.join(self._dataset_loc, self._dataset_name + self._suffix_train))
+            loaded_dts_test = load_from_tsfile_to_dataframe(
+                os.path.join(self._dataset_loc, self._dataset_name + self._suffix_test))
 
             data_train = loaded_dts_train[0]
             y_train = loaded_dts_train[1]
@@ -64,13 +67,14 @@ class DatasetHDD:
             data_test = loaded_dts_test[0]
             y_test = loaded_dts_test[1]
 
-            #concatenate the two dataframes
+            # concatenate the two dataframes
             data_train[self._target] = y_train
             data_test[self._target] = y_test
 
-            data = pd.concat([data_train,data_test], axis=0, keys=['train','test']).reset_index(level=1, drop=True)
+            data = pd.concat([data_train, data_test], axis=0, keys=['train', 'test']).reset_index(level=1, drop=True)
 
             return data
+
 
 class DatasetRAM:
     def __init__(self, dataset, dataset_name):
@@ -97,7 +101,7 @@ class DatasetRAM:
             Name of the dataset
         """
         return self._dataset_name
-    
+
     def load(self):
         """
         Returns
@@ -107,10 +111,12 @@ class DatasetRAM:
         """
         return self._dataset
 
+
 class DatasetLoadFromDir:
     """
     Loads all datasets in a root directory
     """
+
     def __init__(self, root_dir):
         """
         Parameters
@@ -119,7 +125,7 @@ class DatasetLoadFromDir:
             Root directory where the datasets are located
         """
         self._root_dir = root_dir
-    
+
     def load_datasets(self, train_test_exists=True):
         """
         Parameters
@@ -136,16 +142,18 @@ class DatasetLoadFromDir:
 
         data = []
         for dts in datasets:
-            dts = DatasetHDD(dataset_loc=os.path.join(self._root_dir, dts), dataset_name=dts, train_test_exists=train_test_exists)
+            dts = DatasetHDD(dataset_loc=os.path.join(self._root_dir, dts), dataset_name=dts,
+                             train_test_exists=train_test_exists)
             data.append(dts)
         return data
-    
+
+
 class Result:
     """
     Used for passing results to the analyse results class
     """
 
-    def __init__(self,dataset_name, strategy_name, y_true, y_pred):
+    def __init__(self, dataset_name, strategy_name, y_true, y_pred):
         """
         Parameters
         ----------
@@ -172,7 +180,7 @@ class Result:
             Name of the dataset
         """
         return self._dataset_name
-    
+
     @property
     def strategy_name(self):
         """
@@ -182,7 +190,7 @@ class Result:
             Name of the strategy
         """
         return self._strategy_name
-    
+
     @property
     def y_true(self):
         """
@@ -203,24 +211,29 @@ class Result:
         """
         return self._y_pred
 
+
 class SKTimeResult(ABC):
     @abstractmethod
     def save(self):
         """
         Saves the result
         """
+
     @abstractmethod
     def load(self):
         """
         method for loading the results
         """
+
+
 class ResultRAM(SKTimeResult):
     """
     Class for storing the results of the experiments in memory
     """
+
     def __init__(self):
         self._results = []
-    
+
     def save(self, dataset_name, strategy_name, y_true, y_pred, cv_fold):
         """
         Parameters
@@ -247,11 +260,13 @@ class ResultRAM(SKTimeResult):
             sktime results
         """
         return self._results
-        
+
+
 class ResultHDD(SKTimeResult):
     """
     Class for storing the results of the orchestrator
     """
+
     def __init__(self, results_save_dir, strategies_save_dir):
         """
         Parameters
@@ -274,7 +289,7 @@ class ResultHDD(SKTimeResult):
             Path where the strategies will be saved
         """
         return self._strategies_save_dir
-        
+
     def save(self, dataset_name, strategy_name, y_true, y_pred, cv_fold):
         """
         Parameters
@@ -292,8 +307,8 @@ class ResultHDD(SKTimeResult):
         """
         if not os.path.exists(self._results_save_dir):
             os.makedirs(self._results_save_dir)
-        #TODO BUG: write write_results_to_uea_format does not write the results property unless the probas are provided as well.
-        #Dummy probas to make the write_results_to_uea_format function work
+        # TODO BUG: write write_results_to_uea_format does not write the results property unless the probas are provided as well.
+        # Dummy probas to make the write_results_to_uea_format function work
         y_true = list(map(int, y_true))
         y_pred = list(map(int, y_pred))
         num_class_true = np.max(y_true)
@@ -308,10 +323,9 @@ class ResultHDD(SKTimeResult):
                                     dataset_name=dataset_name,
                                     actual_class_vals=y_true,
                                     predicted_class_vals=y_pred,
-                                    actual_probas = probas,
+                                    actual_probas=probas,
                                     resample_seed=cv_fold)
 
-    
     def load(self):
         """
         Returns
@@ -320,12 +334,12 @@ class ResultHDD(SKTimeResult):
             sktime results
         """
         results = []
-        for r,d,f in os.walk(self._results_save_dir):
+        for r, d, f in os.walk(self._results_save_dir):
             for file_to_load in f:
                 if file_to_load.endswith(".csv"):
-                    #found .csv file. Load it and create results object
+                    # found .csv file. Load it and create results object
                     path_to_load = os.path.join(r, file_to_load)
-                    current_row= 0
+                    current_row = 0
                     strategy_name = ""
                     dataset_name = ""
                     y_true = []
@@ -336,15 +350,16 @@ class ResultHDD(SKTimeResult):
                             if current_row == 0:
                                 strategy_name = row[0]
                                 dataset_name = row[1]
-                                current_row +=1
-                            elif current_row >=4:
+                                current_row += 1
+                            elif current_row >= 4:
                                 y_true.append(row[0])
                                 y_pred.append(row[1])
-                                current_row +=1
+                                current_row += 1
                             else:
-                                current_row +=1
-                    #create result object and append
-                    result = Result(dataset_name=dataset_name, strategy_name=strategy_name, y_true=y_true, y_pred=y_pred)
+                                current_row += 1
+                    # create result object and append
+                    result = Result(dataset_name=dataset_name, strategy_name=strategy_name, y_true=y_true,
+                                    y_pred=y_pred)
                     results.append(result)
-        
+
         return results
