+++ sktime/experiments/losses.py
@@ -1,9 +1,6 @@
-
-
 import collections
-import numpy as np
 import pandas as pd
-import logging
+
 
 class Losses(object):
     """
