+++ sktime/experiments/scores.py
@@ -1,7 +1,9 @@
-from abc import ABC, abstractmethod
-from sklearn.metrics import accuracy_score, mean_squared_error
-from sklearn.metrics import precision_score, recall_score
+from abc import ABC
+from abc import abstractmethod
 import numpy as np
+from sklearn.metrics import accuracy_score, mean_squared_error
+
+
 class SKTimeScore(ABC):
     @abstractmethod
     def calculate(self, y_true, y_pred):
@@ -39,10 +41,12 @@ class SKTimeScore(ABC):
             Returns the result of the metric.
         """
 
+
 class ScoreAccuracy(SKTimeScore):
     """
     Calculates the accuracy between the true and predicted lables.
     """
+
     def __init__(self, round_predictions=True):
         """
         Parameters
@@ -51,6 +55,7 @@ class ScoreAccuracy(SKTimeScore):
             Should the predictions be rounded before claculating the accuracy score. This is useful when the accuracy score is used on outputs produced by regressors.
         """
         self._round_predictions = round_predictions
+
     def calculate(self, y_true, y_pred):
         """
         Main method for performing the calculations.
@@ -69,7 +74,6 @@ class ScoreAccuracy(SKTimeScore):
             The accuracy of the prediction.
         """
 
-        
         if self._round_predictions is True:
             y_pred = np.rint(y_pred)
         return accuracy_score(y_true, y_pred)
@@ -94,12 +98,13 @@ class ScoreAccuracy(SKTimeScore):
         errors = (np.array(y_true) - np.array(y_pred)) ** 2
         errors = np.where(errors > 0, 1, 0)
         n = len(errors)
-        
-        std_score = np.std(errors)/np.sqrt(n) 
+
+        std_score = np.std(errors) / np.sqrt(n)
         sum_score = np.sum(errors)
-        avg_score = sum_score/n
+        avg_score = sum_score / n
 
         return avg_score, std_score
+
 
 class ScoreMSE(SKTimeScore):
     """
@@ -123,7 +128,7 @@ class ScoreMSE(SKTimeScore):
             The mean squared error of the prediction.
         """
         return mean_squared_error(y_true, y_pred)
-    
+
     def calculate_per_dataset(self, y_true, y_pred):
         """
         Calculates the loss per dataset
@@ -142,9 +147,9 @@ class ScoreMSE(SKTimeScore):
         """
         errors = (y_true - y_pred) ** 2
         n = len(errors)
-        
-        std_score = np.std(errors)/np.sqrt(n) 
+
+        std_score = np.std(errors) / np.sqrt(n)
         sum_score = np.sum(errors)
-        avg_score = sum_score/n
+        avg_score = sum_score / n
 
         return avg_score, std_score
