+++ sktime/forecasters/__init__.py
@@ -1,0 +1,4 @@
+from .forecasters import ARIMAForecaster
+from .forecasters import ExpSmoothingForecaster
+from .forecasters import DummyForecaster
+from .forecasters import EnsembleForecaster
