+++ sktime/forecasters/tests/test_forecasters.py
@@ -4,9 +4,9 @@ import pandas as pd
 from numpy.testing import assert_array_equal
 from sklearn.metrics import mean_squared_error
 
-from sktime.forecasting.forecasters import DummyForecaster
-from sktime.forecasting.forecasters import ExpSmoothingForecaster
-from sktime.forecasting.forecasters import ARIMAForecaster
+from sktime.forecasters import DummyForecaster
+from sktime.forecasters import ExpSmoothingForecaster
+from sktime.forecasters import ARIMAForecaster
 from sktime.datasets import load_shampoo_sales
 
 __author__ = "Markus Löning"
@@ -25,7 +25,7 @@ FHS = (None, [1], [1, 3], np.array([1]), np.array([1, 3]), np.arange(5))
 y = load_shampoo_sales()
 
 
-# test default forecasters output for different forecasting horizons
+# test default forecasters output for different forecasters horizons
 @pytest.mark.filterwarnings('ignore::FutureWarning')
 @pytest.mark.parametrize("forecaster", FORECASTERS)
 @pytest.mark.parametrize("fh", FHS)
